requirements.txt
celery==4.4.0
kombu==4.6.11
python-dotenv==0.19.1
